src/pyproject.toml
[tool.poetry]
name = "Ripple Project API"
version = "0.1.0"
description = "This is Backend Fastapi App for Ripple Project."
authors = ["Nwolu David"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.111.0"
uvicorn = "^0.30.1"
pydantic = "^2.7.2"
firebase-admin = "^6.5.0"
pyrebase4 = "^4.8.0"
setuptools = "^70.0.0"
requests = "^2.32.3"
pymongo = {extras = ["srv"], version = "^4.7.2"}
pydantic-settings = "^2.3.0"
python-dotenv = "^1.0.1"
logfire = "^0.42.0"
ruff = "^0.4.8"
odmantic = "^1.0.2"
tenacity = "^8.3.0"
passlib = {extras = ["argon2", "bcrypt"], version = "^1.7.4"}
motor = "^3.4.0"
python-jose = "^3.3.0"
emails = "^0.6"
authlib = "^1.3.1"
itsdangerous = "^2.2.0"
python-multipart = "^0.0.9"
pydantic-extra-types = "^2.8.2"
phonenumbers = "^8.13.39"
python-bsonjs = "^0.4.0"
boto3 = "^1.34.136"
jinja2 = "^3.1.4"

[tool.poetry.dev-dependencies]
mypy = "^0.961"
black = "^19.10b0"
isort = "^4.3.21"
autoflake = "^1.3.1"
flake8 = "^3.7.9"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
